+++ packages/state-variable-filter/src/filter.test.ts
@@ -1,28 +1,24 @@
-import { SVFilter } from "./dsp";
+import { createFilter, SvfType } from "./dsp";
 
 describe("SVFilter", () => {
   it("filters the signal", () => {
-    const filter = SVFilter(20);
+    const filter = createFilter(20);
     const input = new Float32Array(20);
     const output = new Float32Array(20);
+    const frequency = new Float32Array(20);
+    const type = SvfType.LowPass;
+    const q = 0.5;
 
     for (let i = 0; i < input.length; i++) {
       input[i] = i % 2 ? 1 : -1;
     }
     expect(input).toMatchSnapshot();
-    filter.update({
-      type: [1],
-      frequency: [10],
-      resonance: [0.5],
-    });
-    filter.fill(input, output);
+    frequency.fill(10);
+    filter(input, output, type, frequency, q);
     expect(output).toMatchSnapshot();
-    filter.update({
-      type: [1],
-      frequency: [5],
-      resonance: [0.5],
-    });
-    filter.fill(input, output);
+
+    frequency.fill(5);
+    filter(input, output, type, frequency, q);
     expect(output).toMatchSnapshot();
   });
 });