+++ packages/synthlet/src/synthlet.ts
@@ -14,11 +14,7 @@ import {
   ParamWorkletNode,
 } from "@synthlet/param";
 import { PolyblepOscillator } from "@synthlet/polyblep-oscillator";
-import {
-  StateVariableFilter,
-  StateVariableFilterInputs,
-  StateVariableFilterType,
-} from "@synthlet/state-variable-filter";
+import { Svf, SvfInputs, SvfType } from "@synthlet/state-variable-filter";
 import { WavetableOscillator } from "@synthlet/wavetable-oscillator";
 import { disposable, Disposable, ParamInput } from "./_worklet";
 import {
@@ -61,7 +57,7 @@ function createSynthlet(context: AudioContext) {
   const lfo = operator(Lfo);
   const noise = operator(Noise);
   const polyblep = operator(PolyblepOscillator);
-  const svf = operator(StateVariableFilter);
+  const svf = operator(Svf);
   const param = operator(Param);
   const wt = operator(WavetableOscillator);
   const clip = operator(ClipAmp);
@@ -153,20 +149,12 @@ function createSynthlet(context: AudioContext) {
     }),
     polyblep: Object.assign(polyblep, {}),
     svf: Object.assign(svf, {
-      lp: (
-        frequency: ParamInput,
-        params?: Partial<StateVariableFilterInputs>
-      ) => svf({ type: StateVariableFilterType.LowPass, frequency, ...params }),
-      hp: (
-        frequency: ParamInput,
-        params?: Partial<StateVariableFilterInputs>
-      ) =>
-        svf({ type: StateVariableFilterType.HighPass, frequency, ...params }),
-      bp: (
-        frequency: ParamInput,
-        params?: Partial<StateVariableFilterInputs>
-      ) =>
-        svf({ type: StateVariableFilterType.BandPass, frequency, ...params }),
+      lp: (frequency: ParamInput, params?: Partial<SvfInputs>) =>
+        svf({ type: SvfType.LowPass, frequency, ...params }),
+      hp: (frequency: ParamInput, params?: Partial<SvfInputs>) =>
+        svf({ type: SvfType.HighPass, frequency, ...params }),
+      bp: (frequency: ParamInput, params?: Partial<SvfInputs>) =>
+        svf({ type: SvfType.BandPass, frequency, ...params }),
     }),
 
     env: Object.assign(adsrEnv, {