+++ packages/state-variable-filter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @synthlet/state-variable-filter
 
+## 0.2.0
+
+- State Variable Filter improvements:
+
+- It uses a better algorithm (by Andrew Simper and Freq Anton Corvest)
+- `frequency` is now a a-rate parameter suitable for modulation
+- Renamed `resonance` to `Q` to match Web Audio API standard
+
 ## 0.1.0
 
 - Initial release